+++ apps/ecommerce/frontend/src/components/MissingResourceCard.tsx
@@ -1,20 +1,10 @@
-import {
-  Badge,
-  Box,
-  Card,
-  Collapse,
-  CopyButton,
-  Flex,
-  IconButton,
-  Menu,
-  Text,
-  Tooltip,
-} from '@contentful/f36-components';
-import { CloseTrimmedIcon, MoreHorizontalIcon } from '@contentful/f36-icons';
+import { Badge, Box, Card, Flex, Text } from '@contentful/f36-components';
 import tokens from '@contentful/f36-tokens';
 import { RenderDragFn } from '@contentful/field-editor-reference/dist/types';
 import { useState } from 'react';
 import { ExternalResourceLink } from '../types';
+import ResourceCardRawData from './ResourceCardRawData';
+import ResourceCardMenu from './ResourceCardMenu';
 
 interface MissingResourceCardProps {
   onRemove: Function;
@@ -23,15 +13,17 @@ interface MissingResourceCardProps {
   errorMessage?: string;
   errorStatus?: number;
   index?: number;
+  total?: number;
   isLoading?: boolean;
   value: string;
+  onMoveToBottom?: Function;
+  onMoveToTop?: Function;
 }
 
 const MissingResourceCard = (props: MissingResourceCardProps) => {
   const [showJson, setShowJson] = useState<boolean>(false);
   const resourceLink = JSON.parse(props.value) as ExternalResourceLink;
-  const resourceProvider = resourceLink.sys.provider;
-  const resourceType = resourceLink.sys.linkType.split(':')[1];
+  const [resourceProvider, resourceType] = resourceLink.sys.linkType?.split(':');
 
   return (
     <Card
@@ -49,22 +41,19 @@ const MissingResourceCard = (props: MissingResourceCardProps) => {
             <Badge variant={props.errorStatus === 404 ? 'warning' : 'negative'}>
               {props.errorMessage}
             </Badge>
-            <Menu offset={[-5, 0]}>
-              <Menu.Trigger>
-                <IconButton icon={<MoreHorizontalIcon />} aria-label="Actions" />
-              </Menu.Trigger>
-              <Menu.List>
-                <Menu.Item onClick={() => props.onRemove(props.index)}>Remove</Menu.Item>
-                <Menu.Divider />
-                <Menu.Item onClick={() => setShowJson(!showJson)}>
-                  {showJson ? 'Hide Data' : 'Show Data'}
-                </Menu.Item>
-              </Menu.List>
-            </Menu>
+            <ResourceCardMenu
+              onRemove={() => props.onRemove(props.index)}
+              isDataVisible={showJson}
+              onToggleDataVisible={() => setShowJson((previousState) => !previousState)}
+              index={props.index}
+              total={props.total}
+              onMoveToBottom={() => props.onMoveToBottom?.call(null, props.index)}
+              onMoveToTop={() => props.onMoveToTop?.call(null, props.index)}
+            />
           </Flex>
         </Flex>
       </Box>
-      <Box padding="spacingXs" paddingLeft="spacingM" paddingRight="spacingM">
+      <Box padding="spacingM">
         <Text
           fontSize="fontSizeL"
           fontWeight="fontWeightDemiBold"
@@ -72,39 +61,11 @@ const MissingResourceCard = (props: MissingResourceCardProps) => {
           isWordBreak={true}>
           Resource is missing or inaccessible
         </Text>
-        <Collapse isExpanded={showJson}>
-          <Flex alignItems="flex-start" fullWidth={true} justifyContent="space-between">
-            <pre
-              style={{
-                position: 'relative',
-                width: '100%',
-                margin: `${tokens.spacingXs} 0`,
-                border: `1px dashed ${tokens.gray300}`,
-                backgroundColor: tokens.gray100,
-                padding: `${tokens.spacingS} 56px ${tokens.spacingS} ${tokens.spacingM}`,
-              }}>
-              {showJson && (
-                <div
-                  style={{ position: 'absolute', right: tokens.spacingXs, top: tokens.spacingXs }}>
-                  <Tooltip content="Hide Data" usePortal={true}>
-                    <IconButton
-                      size="small"
-                      icon={<CloseTrimmedIcon />}
-                      onClick={() => setShowJson(false)}
-                      aria-label="Hide Data"
-                      variant="transparent"
-                    />
-                  </Tooltip>
-                </div>
-              )}
-              <div
-                style={{ position: 'absolute', right: tokens.spacingXs, bottom: tokens.spacingXs }}>
-                <CopyButton value={JSON.stringify(JSON.parse(props.value), null, 2)} />
-              </div>
-              <code>{JSON.stringify(JSON.parse(props.value), null, 2)}</code>
-            </pre>
-          </Flex>
-        </Collapse>
+        <ResourceCardRawData
+          value={props.value}
+          isVisible={showJson}
+          onHide={() => setShowJson(false)}
+        />
       </Box>
     </Card>
   );