+++ apps/ecommerce/frontend/src/locations/ConfigScreen.tsx
@@ -40,7 +40,6 @@ const ConfigScreen = () => {
   const ping = async () => {
     const url = new URL(`http://localhost:8080/dev/api/ping`);
     fetchWithSignedRequest(url, sdk.ids.app, cma, 'GET', {
-      // TODO: consider how we want to pass the provider – path param? query param? header?
       'X-Contentful-Data-Provider': sdk.parameters.instance.provider,
     })
       .then((res) => {