+++ apps/ecommerce/lambda/src/controllers/apiController.ts
@@ -1,13 +1,7 @@
 import { NextFunction, Request, Response } from 'express';
 import axios, { AxiosError, AxiosResponse } from 'axios';
-import { HydratedResourceData, ExternalResourceLink } from '@/src/types';
+import { ExternalResource, ExternalResourceLink } from '@/src/types';
 import { config } from '../config';
-const BASE_URL = `${config.baseUrl}${config.stage === 'prod' ? '' : `/${config.stage}`}`;
-
-const PROVIDERS: Record<string, string> = {
-  shopify: `${BASE_URL}/shopify/resource`,
-  magento: `${BASE_URL}/magento/resource`,
-};
 
 const ApiController = {
   ping: (req: Request, res: Response) => {
@@ -16,25 +10,38 @@ const ApiController = {
 
   resource: async (
     req: Request,
-    res: Response<HydratedResourceData | { error?: unknown; message: string }>,
+    res: Response<ExternalResource | { error?: unknown; message: string }>,
     next: NextFunction
   ) => {
     try {
       const resourceLink: ExternalResourceLink = req.body;
-      const proxyUrl = PROVIDERS[resourceLink.sys.provider.toLowerCase()];
+      const resourceProvider = req.header('x-contentful-data-provider');
+      const proxyBaseUrl = req.header('x-contentful-data-provider-baseurl');
 
-      if (!proxyUrl) {
+      if (!proxyBaseUrl) {
         return res.status(404).send({
           status: 'error',
-          message: `Provider${resourceLink.sys.provider ? `: ${resourceLink.sys.provider}` : ''
+          message: `Base URL configuration for provider${resourceProvider ? `: "${resourceProvider}"` : ''
             } not found`,
         });
+      }
+
+      // TODO: fetch provider config from `${proxyBaseUrl}/config.json`
+      const proxyResourceUrl = new URL(proxyBaseUrl);
+      proxyResourceUrl.pathname += `/resource`;
+
+      // prevent SSRF exploits
+      // TODO: refactor this to get the proxy resource url from the app definition
+      // configuration parameters once this is an internal Contentful service
+      if (proxyResourceUrl.hostname !== config.domainName) {
+        // temporary: only support data provider proxying on the same domain to avoid SSRF exploits
+        throw new Error(`Invalid data provider base URL. Hostname must be ${config.domainName}`);
       }
 
       try {
         let response;
         try {
-          response = await axios.post(proxyUrl, resourceLink);
+          response = await axios.post(proxyResourceUrl.toString(), resourceLink);
         } catch (error) {
           response = (error as AxiosError).response;
         } finally {