+++ apps/ecommerce/lambda/src/routers/apiRouter.spec.ts
@@ -4,20 +4,18 @@ import app from '../app';
 import * as NodeAppsToolkit from '@contentful/node-apps-toolkit';
 import { mockResourceLink } from '../mocks/resourceLink.mock';
 import nock, { RequestBodyMatcher } from 'nock';
-import { mockCombinedResource } from '../mocks/combinedResource.mock';
 import { ExternalResourceLink } from '../types';
 import Sinon from 'sinon';
-
-const BASE_URL = process.env.BASE_URL;
+import { mockExternalResource } from '../mocks/resourceData.mock';
 
 const sandbox = Sinon.createSandbox();
 chai.use(chaiHttp);
 chai.should();
 
 function nockLocalRequest(body: nock.RequestBodyMatcher | ExternalResourceLink) {
-  return nock(`${BASE_URL}`)
+  return nock('https://ecommerce-test.ctfapps.net/test')
     .post(/shopify\/resource$/, <RequestBodyMatcher>body)
-    .reply(200, mockCombinedResource);
+    .reply(200, mockExternalResource);
 }
 
 describe('API Controller', () => {
@@ -37,6 +35,10 @@ describe('API Controller', () => {
         .request(app)
         .get('/api/ping')
         .set('X-Contentful-Data-Provider', 'shopify')
+        .set(
+          'X-Contentful-Data-Provider-BaseURL',
+          'https://ecommerce-test.ctfapps.net/test/shopify'
+        )
         .end((error, res) => {
           expect(res).to.have.status(200);
           expect(res.body).to.have.property('message');
@@ -53,24 +55,14 @@ describe('API Controller', () => {
         .request(app)
         .post('/api/resource')
         .set('X-Contentful-Data-Provider', 'shopify')
+        .set(
+          'X-Contentful-Data-Provider-BaseURL',
+          'https://ecommerce-test.ctfapps.net/test/shopify'
+        )
         .send(mockResourceLink)
         .end((error, res) => {
           expect(res).to.have.status(200);
           expect(res.body).to.haveOwnProperty('description');
-          done();
-        });
-    });
-
-    it('should throw an error when no provider is provided', (done) => {
-      const { provider, ...newResourceLink } = mockResourceLink.sys; // eslint-disable-line @typescript-eslint/no-unused-vars
-
-      chai
-        .request(app)
-        .post('/api/resource')
-        .set('X-Contentful-Data-Provider', 'shopify')
-        .send({ sys: newResourceLink })
-        .end((error, res) => {
-          expect(res).to.have.status(500);
           done();
         });
     });